title code

+++ src/ui/GUI.cs
@@ -173,7 +173,7 @@ public class GUI {
         var yCount = Game.height / size;
 
         var texCoords = new Vector2(0, 0);
-        var texCoordsMax = new Vector2((float)xCount, (float)yCount);
+        var texCoordsMax = new Vector2(xCount, yCount);
         tb.DrawRaw(Game.textureManager.background,
             new VertexColorTexture(new Vector3(left, top, 0), Color4b.Gray, new Vector2(texCoords.X, texCoords.Y)),
             new VertexColorTexture(new Vector3(right, top, 0), Color4b.Gray, new Vector2(texCoordsMax.X, texCoords.Y)),
